moved GoResult to src/error/go.rs
