+++ src/error/mod.rs
@@ -1,5 +1,7 @@
+mod go;
 mod rust;
 
+pub use go::GoResult;
 pub use rust::{clear_error, handle_c_error, set_error, EmptyArg, Error, Panic, Utf8Err, WasmErr};
 
 pub(crate) use rust::empty_err;
